@@ -1,15 +1,27 @@
-// swift-tools-version:4.2
+// swift-tools-version:5.0
 import PackageDescription
 
 let package = Package(
     name: "Nimble",
+    platforms: [
+      .macOS(.v10_10), .iOS(.v8), .tvOS(.v9)
+    ],
     products: [
         .library(name: "Nimble", targets: ["Nimble"]),
+    ],
+    dependencies: [
+        .package(url: "https://github.com/mattgallagher/CwlPreconditionTesting.git", .exact("1.2.0")),
     ],
     targets: [
         .target(
             name: "Nimble", 
-            dependencies: []
+            dependencies: {
+                #if os(macOS)
+                return ["CwlPreconditionTesting"]
+                #else
+                return []
+                #endif
+            }()
         ),
         .testTarget(
             name: "NimbleTests", 
@@ -17,5 +29,5 @@ let package = Package(
             exclude: ["objc"]
         ),
     ],
-    swiftLanguageVersions: [.v4_2]
+    swiftLanguageVersions: [.v5]
 )
