@@ -3,11 +3,7 @@
 #import "NMBStringify.h"
 #import "DSL.h"
 
-#if TARGET_OS_TV
-    #import "CwlPreconditionTesting_POSIX.h"
-#else
-    #import "CwlPreconditionTesting.h"
-#endif
+#import "CwlPreconditionTesting.h"
 
 FOUNDATION_EXPORT double NimbleVersionNumber;
 FOUNDATION_EXPORT const unsigned char NimbleVersionString[];
