@@ -4,9 +4,13 @@
 
 open class Behavior<Context> {
 
-    public static var name: String { return String(describing: self) }
     /**
-     override this method in your behavior to define a set of reusable examples.
+     Override this variable if you want to provide custom name for this example group.
+    */
+    open class var name: String { return String(describing: self) }
+
+    /**
+     Override this method in your behavior to define a set of reusable examples.
 
      This behaves just like an example group defines using `describe` or `context`--it may contain any number of `beforeEach`
      and `afterEach` closures, as well as any number of examples (defined using `it`).
