@@ -1,39 +1,78 @@
 import Foundation
 import XCTest
 
-// NOTE: This file is not intended to be included in the Xcode project or CocoaPods.
-//       It is picked up by the Swift Package Manager during its build process.
-
 #if SWIFT_PACKAGE
 
 open class QuickConfiguration: NSObject {
     open class func configure(_ configuration: Configuration) {}
 }
 
-#if canImport(Darwin)
+#endif
 
-internal func qck_enumerateSubclasses<T: AnyObject>(_ klass: T.Type, block: (T.Type) -> Void) {
-    var classesCount = objc_getClassList(nil, 0)
+extension QuickConfiguration {
+    #if !canImport(Darwin)
+    private static var configurationSubclasses: [QuickConfiguration.Type] = []
+    #endif
 
-    guard classesCount > 0 else {
-        return
+    /// Finds all direct subclasses of QuickConfiguration and passes them to the block provided.
+    /// The classes are iterated over in the order that objc_getClassList returns them.
+    ///
+    /// - parameter block: A block that takes a QuickConfiguration.Type.
+    ///                    This block will be executed once for each subclass of QuickConfiguration.
+    private static func enumerateSubclasses(_ block: (QuickConfiguration.Type) -> Void) {
+        #if canImport(Darwin)
+        var classesCount = objc_getClassList(nil, 0)
+
+        guard classesCount > 0 else {
+            return
+        }
+
+        let classes = UnsafeMutablePointer<AnyClass?>.allocate(capacity: Int(classesCount))
+        defer { free(classes) }
+
+        classesCount = objc_getClassList(AutoreleasingUnsafeMutablePointer(classes), classesCount)
+
+        var configurationSubclasses: [QuickConfiguration.Type] = []
+        for i in 0..<classesCount {
+            guard
+                let subclass = classes[Int(i)],
+                let superclass = class_getSuperclass(subclass),
+                superclass == QuickConfiguration.self
+                else { continue }
+
+            // swiftlint:disable:next force_cast
+            configurationSubclasses.append(subclass as! QuickConfiguration.Type)
+        }
+        #endif
+
+        configurationSubclasses.forEach(block)
     }
 
-    let classes = UnsafeMutablePointer<AnyClass?>.allocate(capacity: Int(classesCount))
-    classesCount = objc_getClassList(AutoreleasingUnsafeMutablePointer(classes), classesCount)
-
-    var subclass: AnyClass!
-    for i in 0..<classesCount {
-        subclass = classes[Int(i)]
-
-        guard let superclass = class_getSuperclass(subclass), superclass == klass else { continue }
-
-        block(subclass as! T.Type) // swiftlint:disable:this force_cast
+    #if canImport(Darwin)
+    @objc
+    static func configureSubclassesIfNeeded(world: World) {
+        _configureSubclassesIfNeeded(world: world)
     }
+    #else
+    static func configureSubclassesIfNeeded(_ configurationSubclasses: [QuickConfiguration.Type]? = nil, world: World) {
+        // Storing subclasses for later use (will be used when running additional test suites)
+        if let configurationSubclasses = configurationSubclasses {
+            self.configurationSubclasses = configurationSubclasses
+        }
 
-    free(classes)
+        _configureSubclassesIfNeeded(world: world)
+    }
+    #endif
+
+    private static func _configureSubclassesIfNeeded(world: World) {
+        if world.isConfigurationFinalized { return }
+
+        // Perform all configurations (ensures that shared examples have been discovered)
+        world.configure { configuration in
+            enumerateSubclasses { configurationClass in
+                configurationClass.configure(configuration)
+            }
+        }
+        world.finalizeConfiguration()
+    }
 }
-
-#endif
-
-#endif
