@@ -48,36 +48,22 @@ open class QuickSpec: QuickSpecBase {
     /// discovered powered by Objective-C runtime), so we needed the alternative
     /// way.
     override open class var defaultTestSuite: XCTestSuite {
-        configureDefaultTestSuite()
+        QuickConfiguration.configureSubclassesIfNeeded(world: World.sharedWorld)
+
+        // Let's gather examples for each spec classes. This has the same effect
+        // as listing spec classes in `LinuxMain.swift` on Linux.
+        gatherExamplesIfNeeded()
 
         return super.defaultTestSuite
     }
 
-    private static func configureDefaultTestSuite() {
-        let world = World.sharedWorld
-
-        if !world.isConfigurationFinalized {
-            // Perform all configurations (ensures that shared examples have been discovered)
-            world.configure { configuration in
-                qck_enumerateSubclasses(QuickConfiguration.self) { configurationClass in
-                    configurationClass.configure(configuration)
-                }
-            }
-            world.finalizeConfiguration()
-        }
-
-        // Let's gather examples for each spec classes. This has the same effect
-        // as listing spec classes in `LinuxMain.swift` on Linux.
-        _ = allTests
-    }
-
-    override open class func _qck_testMethodSelectors() -> [_QuickSelectorWrapper] {
-        let examples = World.sharedWorld.examples(self)
+    override open class func _qck_testMethodSelectors() -> [String] {
+        let examples = World.sharedWorld.examples(forSpecClass: self)
 
         var selectorNames = Set<String>()
         return examples.map { example in
             let selector = addInstanceMethod(for: example, classSelectorNames: &selectorNames)
-            return _QuickSelectorWrapper(selector: selector)
+            return NSStringFromSelector(selector)
         }
     }
 
@@ -106,16 +92,11 @@ open class QuickSpec: QuickSpecBase {
     }
 #endif
 
-    static var allTestsCache = [String: [(String, (QuickSpec) -> () throws -> Void)]]()
-
+#if !canImport(Darwin)
     public class var allTests: [(String, (QuickSpec) -> () throws -> Void)] {
-        if let cached = allTestsCache[String(describing: self)] {
-            return cached
-        }
-
         gatherExamplesIfNeeded()
 
-        let examples = World.sharedWorld.examples(self)
+        let examples = World.sharedWorld.examples(forSpecClass: self)
         let result = examples.map { example -> (String, (QuickSpec) -> () throws -> Void) in
             return (example.name, { spec in
                 return {
@@ -124,17 +105,19 @@ open class QuickSpec: QuickSpecBase {
                 }
             })
         }
-        allTestsCache[String(describing: self)] = result
         return result
     }
+#endif
 
     internal static func gatherExamplesIfNeeded() {
         let world = World.sharedWorld
-        let rootExampleGroup = world.rootExampleGroupForSpecClass(self)
-        if rootExampleGroup.examples.isEmpty {
-            world.currentExampleGroup =  rootExampleGroup
+        let rootExampleGroup = world.rootExampleGroup(forSpecClass: self)
+        guard rootExampleGroup.examples.isEmpty else {
+            return
+        }
+
+        world.performWithCurrentExampleGroup(rootExampleGroup) {
             self.init().spec()
-            world.currentExampleGroup = nil
         }
     }
 }
