@@ -1,22 +1,5 @@
 #import "QuickSpecBase.h"
 
-#pragma mark - _QuickSelectorWrapper
-
-@interface _QuickSelectorWrapper ()
-@property(nonatomic, assign) SEL selector;
-@end
-
-@implementation _QuickSelectorWrapper
-
-- (instancetype)initWithSelector:(SEL)selector {
-    self = [super init];
-    _selector = selector;
-    return self;
-}
-
-@end
-
-
 #pragma mark - _QuickSpecBase
 
 @implementation _QuickSpecBase
@@ -33,11 +16,11 @@
  @return An array of invocations that execute the newly defined example methods.
  */
 + (NSArray<NSInvocation *> *)testInvocations {
-    NSArray<_QuickSelectorWrapper *> *wrappers = [self _qck_testMethodSelectors];
-    NSMutableArray<NSInvocation *> *invocations = [NSMutableArray arrayWithCapacity:wrappers.count];
+    NSArray<NSString *> *selectors = [self _qck_testMethodSelectors];
+    NSMutableArray<NSInvocation *> *invocations = [NSMutableArray arrayWithCapacity:selectors.count];
 
-    for (_QuickSelectorWrapper *wrapper in wrappers) {
-        SEL selector = wrapper.selector;
+    for (NSString *selectorString in selectors) {
+        SEL selector = NSSelectorFromString(selectorString);
         NSMethodSignature *signature = [self instanceMethodSignatureForSelector:selector];
         NSInvocation *invocation = [NSInvocation invocationWithMethodSignature:signature];
         invocation.selector = selector;
@@ -48,7 +31,7 @@
     return invocations;
 }
 
-+ (NSArray<_QuickSelectorWrapper *> *)_qck_testMethodSelectors {
++ (NSArray<NSString *> *)_qck_testMethodSelectors {
     return @[];
 }
 
